Refactor validaton and sanitizing handling simplified
Validator and sanitizer no longer raise errors when invalid, but are rather
replace with a default identity function in all cases except when the user
provided option is a callable. Sanitizers only receive the value, with the
idea to encourage general purpose sanitizers, such as whitespace trimming,
length constraints etc, leaving more specialized logic for validators.

Validators will in addition to the field value also receive the field
attributes as a param. With this extra information validity can be checked
with higher precision. This is also useful for adding error messages to
be displayed on the next page.

These changes greatly simplify the piece of code in question and also
makes sense of why there are both sanitizers and validators available.